Move getDisputeRevertMessage method
