+++ eth/dispute_transition.go
@@ -1,13 +1,11 @@
 package eth
 
 import (
-	"context"
 	"math/big"
 
 	"github.com/Worldcoin/hubble-commander/contracts/rollup"
 	"github.com/Worldcoin/hubble-commander/eth/chain"
 	"github.com/Worldcoin/hubble-commander/models"
-	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/pkg/errors"
@@ -128,20 +126,6 @@ func (c *Client) waitForDispute(batchID *models.Uint256, batchHash *common.Hash,
 		return NewDisputeTxRevertedError(batchID.Uint64(), err.Error())
 	}
 	return NewUnknownDisputeTxRevertedError(batchID.Uint64())
-}
-
-func (c *Client) getDisputeRevertMessage(tx *types.Transaction, txReceipt *types.Receipt) error {
-	callMsg := ethereum.CallMsg{
-		From:     c.Blockchain.GetAccount().From,
-		To:       tx.To(),
-		Gas:      tx.Gas(),
-		GasPrice: tx.GasPrice(),
-		Value:    tx.Value(),
-		Data:     tx.Data(),
-	}
-
-	_, err := c.Blockchain.GetBackend().CallContract(context.Background(), callMsg, txReceipt.BlockNumber)
-	return err
 }
 
 func (c *Client) isBatchAlreadyDisputed(batchID *models.Uint256, batchHash *common.Hash) error {
